refactor: create utility module
