+++ src/service/config.rs
@@ -92,7 +92,7 @@ impl Config {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::service::test_util::testdir_and_path;
+    use crate::util::test::testdir_and_path;
 
     use tempfile::TempDir;
