+++ src/service/init.rs
@@ -113,7 +113,7 @@ fn remove_path_depth(path: &mut PathBuf, depth: usize) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::service::test_util::{template_dir, testdir_and_path};
+    use crate::util::test::{template_dir, testdir_and_path};
 
     #[test]
     fn directories_exist() {
