+++ src/service/mod.rs
@@ -2,6 +2,3 @@ pub mod build_system;
 pub mod config;
 pub mod init;
 pub mod vcs;
-
-#[cfg(test)]
-pub mod test_util;
