+++ src/service/vcs.rs
@@ -100,7 +100,7 @@ fn git_init(path: &Path) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::service::test_util::testdir_and_path;
+    use crate::util::test::testdir_and_path;
 
     use tempfile::TempDir;
